getter and setter signatures
--HG--
extra : convert_revision : svn%3A45ebfb91-7180-4db0-a2bd-9f34e122093e/trunk%40761

+++ src/lib/binary/bytearray.cc
@@ -139,7 +139,7 @@ JS_METHOD(_concat) {
 	return args.This();
 }
 
-v8::Handle<v8::Value> _get(size_t index, const v8::AccessorInfo &info) {
+v8::Handle<v8::Value> _get(uint32_t index, const v8::AccessorInfo &info) {
 	ByteStorage * bs = BS_OTHER(info.This());
 	size_t len = bs->getLength();
 	if (index < 0 || index >= len) { return v8::Undefined(); }
@@ -147,7 +147,7 @@ v8::Handle<v8::Value> _get(size_t index, const v8::AccessorInfo &info) {
 	return JS_INT(bs->getByte(index));
 }
 
-v8::Handle<v8::Value> _set(size_t index, v8::Local<v8::Value> value, const v8::AccessorInfo &info) {
+v8::Handle<v8::Value> _set(uint32_t index, v8::Local<v8::Value> value, const v8::AccessorInfo &info) {
 	ByteStorage * bs = BS_OTHER(info.This());
 	size_t len = bs->getLength();
 	if (index < 0 || index >= len) { return v8::Undefined(); }
